+++ classes/actions.py
@@ -14,15 +14,13 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with Openplotter. If not, see <http://www.gnu.org/licenses/>.
-import gammu
-import platform
-import subprocess
-import time
+import gammu, platform, subprocess, time, json, requests
 
 import paho.mqtt.publish as publish
 
 from classes.gmailbot import GmailBot
 from classes.twitterbot import TwitterBot
+from check_vessel_self import checkVesselSelf
 
 if platform.machine()[0:3] == 'arm':
 	import RPi.GPIO as GPIO
@@ -34,44 +32,34 @@ class Actions:
 	def __init__(self, SK):
 		self.SK = SK
 		self.conf = SK.conf
+		self.home = SK.home
+		self.currentpath = SK.currentpath
+		self.time_units = [_('no repeat'), _('seconds'), _('minutes'), _('hours'), _('days')]
 		self.options = []
 		# ATENTION. If order changes, edit "run_action()" and ctrl_actions.py
 		# 0 name, 1 message, 2 field data, 3 unique ID
-		self.options.append([_('wait'), _('Enter seconds to wait in the field below.'), 1, 'ACT1'])
+		self.options.append([_('wait'), _('Enter seconds to wait in field "data".'), 1, 'ACT1'])
 		self.options.append([_('command'), _('Enter a Linux command and arguments in the field below.'), 1, 'ACT2'])
-		self.options.append([_('reset'), 0, 0, 'ACT3'])
-		self.options.append([_('shutdown'), 0, 0, 'ACT4'])
+		self.options.append([_('reset system'), 0, 0, 'ACT3'])
+		self.options.append([_('shutdown system'), 0, 0, 'ACT4'])
+		self.options.append([_('startup stop'), 0, 0, 'ACT7'])
+		self.options.append([_('startup restart'), 0, 0, 'ACT8'])
 		self.options.append([_('stop NMEA multiplexer'), 0, 0, 'ACT5'])
 		self.options.append([_('reset NMEA multiplexer'), 0, 0, 'ACT6'])
-		self.options.append([_('stop Signal K server'), 0, 0, 'ACT7'])
-		self.options.append([_('reset Signal K server'), 0, 0, 'ACT8'])
-		self.options.append([_('stop WiFi access point'),
-							 _('Be careful, if you are connected by remote you may not be able to reconnect again.'), 0,
-							 'ACT9'])
-		self.options.append([_('start WiFi access point'),
-							 _('Be sure you have filled in all fields in "WiFi AP" tab and enabled WiFi access point.'),
-							 0, 'ACT10'])
-		self.options.append([_('stop SDR-AIS'), 0, 0, 'ACT11'])
-		self.options.append([_('reset SDR-AIS'), _(
-			'Be sure you have filled in Gain and Correction fields in "SDR-AIS" tab and enabled AIS NMEA generation.'),
-							 0, 'ACT12'])
-		self.options.append([_('publish Twitter'), _(
-			'Be sure you have filled in all fields in "Accounts" tab, and enabled Twitter checkbox.\n\nEnter text to publish in the field below.'),
-							 1, 'ACT13'])
-		self.options.append([_('send e-mail'), _(
-			'Be sure you have filled in all fields in "Accounts" tab, and enabled Gmail checkbox.\n\nEnter the subject in the field below.'),
-							 1, 'ACT14'])
-		self.options.append([_('send SMS'), _(
-			'Be sure you have enabled sending SMS in "SMS" tab.\n\nEnter the text in the field below.'), 1, 'ACT21'])
+		self.options.append([_('stop WiFi access point'),_('Access point will be disabled.\n\nIf you are on a headless system, you will not be able to reconnect again.\n\nAre you sure?'),0,'ACT9'])
+		self.options.append([_('start WiFi access point'),_('Be sure you have filled in all fields in "WiFi AP" tab.'),0, 'ACT10'])
+		#self.options.append([_('stop SDR-AIS'), 0, 0, 'ACT11'])
+		#self.options.append([_('reset SDR-AIS'), _('Be sure you have filled in Gain and Correction fields in "SDR-AIS" tab and enabled AIS NMEA generation.'),0, 'ACT12'])
+		self.options.append([_('publish Twitter'), _('Be sure you have filled in all fields in "Accounts" tab, and enabled Twitter checkbox.\n\nEnter text to publish in the field "data".'),1, 'ACT13'])
+		self.options.append([_('send e-mail'), _('Be sure you have filled in all fields in "Accounts" tab, and enabled Gmail checkbox.\n\nEnter the subject in the field "data".'),1, 'ACT14'])
+		self.options.append([_('send SMS'), _('Be sure you have enabled sending SMS in "SMS" tab.\n\nEnter the text in field "data".'), 1, 'ACT21'])
 		self.options.append([_('play sound'), 'OpenFileDialog', 1, 'ACT15'])
 		self.options.append([_('stop all sounds'), 0, 0, 'ACT16'])
-		self.options.append([_('show message'), _('Enter the message in the field below.'), 1, 'ACT17'])
+		self.options.append([_('show message'), _('Enter the message in field "data".'), 1, 'ACT17'])
 		self.options.append([_('close all messages'), 0, 0, 'ACT18'])
-		self.options.append([_('start all actions'), 0, 0, 'ACT19'])
-		self.options.append([_('stop all actions'), _(
-			'This action will stop all the triggers except the trigger which has an action "start all actions" defined.'),
-							 0, 'ACT20'])
-
+		#self.options.append([_('start all actions'), 0, 0, 'ACT19'])
+		#self.options.append([_('stop all actions'), _('This action will stop all the triggers except the trigger which has an action "start all actions" defined.'),0, 'ACT20'])
+		'''
 		#init GPIO
 		GPIO.setmode(GPIO.BCM)
 		GPIO.setwarnings(False)
@@ -103,19 +91,11 @@ class Actions:
 				self.options.append([_('Publish on topic ') + i[1], 0, 1, i[2]])
 			except Exception, e:
 				print str(e)
-
-		self.time_units = [_('no repeat'), _('seconds'), _('minutes'), _('hours'), _('days')]
-
-		self.home = SK.home
-		self.currentpath = SK.currentpath
+		'''
 
 	def getOptionsListIndex(self, data):
 		for index, item in enumerate(self.options):
 			if item[3] == data: return index
-
-	def getoutlistIndex(self, data):
-		for index, item in enumerate(self.out_list):
-			if item[4] == data: return index
 
 	def getmqttlistIndex(self, data):
 		for index, item in enumerate(self.mqtt_list):
@@ -132,16 +112,15 @@ class Actions:
 		elif option == 'ACT1':
 			try:
 				wait = float(text)
+				time.sleep(wait)
 			except:
-				wait = 1.0
-			time.sleep(wait)
+				print 'Wait action needs a number'
 		elif option == 'ACT2':
 			if text:
 				try:
 					text = text.split(' ')
 					subprocess.Popen(text)
-				except Exception, e:
-					print str(e)
+				except Exception, e: print str(e)
 		elif option == 'ACT3':
 			subprocess.Popen(['sudo', 'reboot'])
 		elif option == 'ACT4':
@@ -152,39 +131,17 @@ class Actions:
 			subprocess.call(['pkill', '-9', 'kplex'])
 			subprocess.Popen('kplex')
 		elif option == 'ACT7':
-			subprocess.Popen(["pkill", '-f', "signalk-server-node"])
+			subprocess.Popen(['startup', 'stop'])
 		elif option == 'ACT8':
-			subprocess.call(["pkill", '-f', "signalk-server-node"])
-			subprocess.Popen(self.home + '/.config/signalk-server-node/bin/openplotter',
-							 cwd=self.home + '/.config/signalk-server-node')
+			subprocess.Popen(['startup', 'restart'])
 		elif option == 'ACT9':
-			subprocess.Popen(['sudo', 'python', self.currentpath + '/wifi_server.py', '0'])
+			subprocess.Popen(['sudo', 'python', self.currentpath +'/wifi_server.py', '0'])
 			conf.set('WIFI', 'enable', '0')
-			conf.read()
 		elif option == 'ACT10':
-			subprocess.Popen(['sudo', 'python', self.currentpath + '/wifi_server.py', '1', wlan, passw2, ssid2])
+			subprocess.Popen(['sudo', 'python', self.currentpath+'/wifi_server.py', '1'])
 			conf.set('WIFI', 'enable', '1')
-			conf.read()
-		elif option == 'ACT11':
-			subprocess.Popen(['pkill', '-9', 'aisdecoder'])
-			subprocess.Popen(['pkill', '-9', 'rtl_fm'])
-			conf.set('AIS-SDR', 'enable', '0')
-			conf.read()
-		elif option == 'ACT12':
-			gain = conf.get('AIS-SDR', 'gain')
-			ppm = conf.get('AIS-SDR', 'ppm')
-			channel = conf.get('AIS-SDR', 'channel')
-			subprocess.call(['pkill', '-9', 'aisdecoder'])
-			subprocess.call(['pkill', '-9', 'rtl_fm'])
-			frecuency = '161975000'
-			if channel == 'b': frecuency = '162025000'
-			rtl_fm = subprocess.Popen(['rtl_fm', '-f', frecuency, '-g', gain, '-p', ppm, '-s', '48k'],
-									  stdout=subprocess.PIPE)
-			aisdecoder = subprocess.Popen(
-				['aisdecoder', '-h', 'localhost', '-p', '10110', '-a', 'file', '-c', 'mono', '-d', '-f', '/dev/stdin'],
-				stdin=rtl_fm.stdout)
-			conf.set('AIS-SDR', 'enable', '1')
-			conf.read()
+		#elif option == 'ACT11':
+		#elif option == 'ACT12':
 		elif option == 'ACT13':
 			now = time.strftime("%H:%M:%S")
 			tweetStr = now + ' ' + text
@@ -196,11 +153,12 @@ class Actions:
 			try:
 				msg = TwitterBot(apiKey, apiSecret, accessToken, accessTokenSecret)
 				msg.send(tweetStr)
-			except Exception, e:
-				print str(e)
+			except Exception, e: print str(e)
 		elif option == 'ACT14':
 			subject = text
 			body = ''
+			for i in self.SK.list_SK:
+				body += i[1]+': '+str(i[2])+_('\nsource: ')+i[0]+_('\ntimestamp: ')+i[7]+'\n\n'
 			GMAIL_USERNAME = conf.get('GMAIL', 'gmail')
 			GMAIL_PASSWORD = conf.get('GMAIL', 'password')
 			recipient = conf.get('GMAIL', 'recipient')
@@ -208,8 +166,7 @@ class Actions:
 			try:
 				msg = GmailBot(GMAIL_USERNAME, GMAIL_PASSWORD, recipient)
 				msg.send(subject, body)
-			except Exception, e:
-				print str(e)
+			except Exception, e: print str(e)
 		elif option == 'ACT15':
 			subprocess.Popen(['mpg123', '-q', text])
 		elif option == 'ACT16':
@@ -233,8 +190,7 @@ class Actions:
 					'Number': conf.get('SMS', 'phone'),
 				}
 				sm.SendSMS(message)
-			except Exception, e:
-				print str(e)
+			except Exception, e: print str(e)
 		elif option[:4] == 'MQTT':
 			topic = self.mqtt_list[self.getmqttlistIndex(option)][1]
 			payload = text